refactor: rename to input / output
